Add files via upload
# 1 - 23 - 2025

## 128. Longest Consecutive Sequence
Description: Makes a set of the nums list and then iterates through the set and if there is no previous sequential value, it then starts counting the consecutive values from that value. Then when the sequence is broken it checks if the current count is longer than the longest count so far and if it is set that count to the longest.
Runtime: O(n)
Space: O(n)
<hr/>

## 15. 3Sum
Description: Sorts the list to make the movement of the pointers more logical. Then it iterates over the list with each element starting from the front and has two pointers, one starting on the left side of the list after the current element and another on the right side from the end of the list. When 3sums are found they are added to the result list. Repeats of numbers are skipped to avoid 3sums with the same values.
Runtime: O(n^2) because of the sorting (O(nlog(n))) and the iterating over the list with pointers O(n) 
Space = O(n) - the copy of the list sorted
<hr/>